wip
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3690 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ lib/qfevaluationresultsmodel.cpp
@@ -758,7 +758,7 @@ void QFEvaluationResultsModel::calcStatistics(QString resultName, double& averag
                 QVector<bool> mask;
                 if (runsel && d.size()==runsel->leaveoutGetRunCount()) {
                     mask=QFRDRRunSelectionsInterface_getRunSelectionAsBoolVec(runsel);
-                } else if (maskrdr && d.size()==maskrdr->maskGetWidth()*maskrdr->maskGetHeight()) {
+                } else if (maskrdr && d.size()==(int64_t)maskrdr->maskGetWidth()*(int64_t)maskrdr->maskGetHeight()) {
                     mask=QFRDRImageMaskInterface_getMaskAsBoolVec(maskrdr);
                 }
                 if (mindRunsForVectorStat && mask.size()>=d.size()){
